# docs/swagger.yaml
swagger: "2.0"
info:
  title: User API
  description: User authentication and profile management API
  version: "1.0.0"
  contact:
    name: API Support

host: localhost:8080
basePath: /
schemes:
  - http
  - https

consumes:
  - application/json
produces:
  - application/json

securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: JWT Authorization header (Bearer token)

paths:
  /login:
    post:
      summary: User Login
      description: Authenticate user with email and password
      operationId: userLogin
      tags:
        - Authentication
      parameters:
        - in: body
          name: body
          description: User login credentials
          required: true
          schema:
            $ref: "#/definitions/LoginRequest"
      responses:
        200:
          description: Login successful
          schema:
            $ref: "#/definitions/LoginResponse"
        400:
          description: Invalid request body
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: Invalid email or password
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /api/profile:
    get:
      summary: Get User Profile
      description: Retrieve authenticated user's profile information
      operationId: getUserProfile
      tags:
        - Profile
      security:
        - Bearer: []
      responses:
        200:
          description: User profile retrieved successfully
          schema:
            $ref: "#/definitions/UserProfile"
        401:
          description: Unauthorized access
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"

definitions:
  LoginRequest:
    type: object
    required:
      - user
    properties:
      user:
        type: object
        required:
          - email
          - password
        properties:
          email:
            type: string
            format: email
            example: "user@example.com"
          password:
            type: string
            format: password
            example: "password123"

  LoginResponse:
    type: object
    properties:
      user:
        type: object
        properties:
          id:
            type: integer
            format: uint
            example: 1
          name:
            type: string
            example: "John Doe"
          email:
            type: string
            format: email
            example: "user@example.com"
          access_token:
            type: string
            example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."

  UserProfile:
    type: object
    properties:
      id:
        type: integer
        format: uint
        example: 1
      name:
        type: string
        example: "John Doe"
      email:
        type: string
        format: email
        example: "user@example.com"
      birthday:
        type: string
        format: date
        example: "1990-01-15"
      current_team:
        $ref: "#/definitions/TeamSummary"
      position:
        $ref: "#/definitions/Position"
      projects:
        type: array
        items:
          $ref: "#/definitions/ProjectSummary"
      skills:
        type: array
        items:
          $ref: "#/definitions/UserSkillSummary"

  TeamSummary:
    type: object
    properties:
      id:
        type: integer
        format: uint
        example: 1
      name:
        type: string
        example: "Backend Team"

  Position:
    type: object
    properties:
      id:
        type: integer
        format: uint
        example: 1
      name:
        type: string
        example: "Senior Developer"
      abbreviation:
        type: string
        example: "SD"

  ProjectSummary:
    type: object
    properties:
      id:
        type: integer
        format: uint
        example: 1
      name:
        type: string
        example: "Project Alpha"
      abbreviation:
        type: string
        example: "PA"
      start_date:
        type: string
        format: date-time
        example: "2024-01-01T00:00:00Z"
      end_date:
        type: string
        format: date-time
        example: "2024-12-31T00:00:00Z"

  UserSkillSummary:
    type: object
    properties:
      id:
        type: integer
        format: uint
        example: 1
      name:
        type: string
        example: "Go Programming"

  ErrorResponse:
    type: object
    properties:
      error:
        type: string
        example: "Invalid email or password"